Suppress the unused ":category" field in syntax checking.
A change in clang-c/Index.h deprecate the function
clang_getDiagnosticCategoryName(...) (since clang 3.1 I think). And
the category is not really important for the moment.

+++ server/TUManager.cpp
@@ -54,7 +54,7 @@ CXTranslationUnit TUManager::parse(const std::string &              filename,
 
       tu = clang_parseTranslationUnit(index_,
                                       filename.c_str(),
-                                      argv, nbArgs,
+                                      argv, static_cast<int>(nbArgs),
                                       0, 0,
                                       clang_defaultEditingTranslationUnitOptions()
                                       | CXTranslationUnit_PrecompiledPreamble
